requirements.txt
Django==4.2.28
werkzeug==3.1.6
# Keep canonical DRF package only; remove legacy django-rest-framework package.
djangorestframework==3.15.2
googlemaps==3.0.2
# Python 3.12 compatible scientific stack (used in apps/api/utils.py)
numpy==1.26.4
scipy==1.13.1
python-decouple==3.1
pytz==2024.1
psycopg2-binary==2.9.9
gunicorn==22.0.0

google-auth==2.35.0

PyJWT[crypto]==2.9.0
msgpack==1.1.0
zstandard==0.23.0
paho-mqtt==1.6.1
